pyproject.toml
[tool.poetry]
name = "langgraph-example"
version = "0.1.0"
description = ""
authors = ["Allan Ferreira <ferrerallan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchain = "^0.3.14"
langchain-openai = "^0.3.0"
langgraph = "^0.2.62"
grandalf = "^0.8"
streamlit = "^1.41.1"
chromadb = "^0.6.2"
sentence-transformers = "^3.3.1"
langchain-community = "^0.3.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
